Allow custom dependency repositories in SCIP build tool
